+++ packages/kernel/src/domain/index.ts
@@ -1,4 +1,3 @@
 export * from "./schema-diff";
 export * from "./snapshot-dsl";
 export * from "./meta-definition-registry";
-export * from "./demo-meta-registry";